requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-environ>=0.11
whitenoise>=6.6
gunicorn>=21.2
PyMySQL>=1.1
tzdata>=2024.1
psycopg[binary]>=3
django-storages[boto3]>=1.14
boto3>=1.34         # 如果用 S3/R2
django-widget-tweaks>=1.5.0
cryptography>=42.0.0
Pillow>=10.0.0
